telecom-lakehouse/requirements.txt
pyspark==3.5.0
delta-spark==3.1.0
great-expectations==0.18.0
pandas==2.1.4
numpy==1.26.0
matplotlib==3.8.0
seaborn==0.13.0
pytest==7.4.0
databricks-sdk==0.20.0
databricks-connect==14.3.0
python-dotenv==1.0.0
pyarrow==14.0.0
